Develop

+++ my-doctor-data/src/main/java/com/mydoctorapp/data/repository/MedicineRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.mydoctorapp.data.bean.Medicine;
 
-public interface MedicineRepository extends JpaRepository<Medicine,Long>{
+public interface MedicineRepository extends JpaRepository<Medicine,Integer>{
 
 }
