+++ my-doctor-data/src/main/java/com/mydoctorapp/data/repository/SymptomRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.mydoctorapp.data.bean.Medicine;
 import com.mydoctorapp.data.bean.Symptom;
 
-public interface SymptomRepository extends JpaRepository<Symptom,Long>{
+public interface SymptomRepository extends JpaRepository<Symptom,Integer>{
 
 }
